API KEY FIX
Blizzard is discontinuing the legacy API servers Feb 15th 2016, but for
some reason they're not working now. Maybe it's due to Blizzcon
happening and is temporary, but either way this fix will have to happen
sooner or later. Might write a proxy API to handle this, not sure if the
key would be overused or not.  This is the fix for now. All users must
obtain their own API key to use the sheet.