Removed duplicated code
Thise changed where done with this in mind:
1. Rework code so it does not repeat itself (this also makes it easier
to read as you can better see what it is doing, and easier to modify as
you in most cases will make 1 change instead of 16)
2.  Make more of the variables inside Objects (or arrays) as they are
the best places to store data.

DO NOT MERGE THIS YET.

I have put some variables inside the comment at the end of my added
code. I wrote down all the changes i made to the varialbes that needs
reworking outside the code i supplied.