bloodrash commas in AP, force retry due to frequent timeouts
github user bloodrash has added in a function to make AP more readable

added a try{ to try to deal with these frequent gateway timeouts. not
sure if the issue is on API's end or sheets. In the cases of bad
names/realms there will not be an infinite loop, the function will
timeout on its own and return as an internal error.