bloodrash commas in AP, force retry due to frequent timeouts
had to resubmitt due to format issue..

added a try{ to try to deal with these frequent gateway timeouts. not
sure if the issue is on API's end or sheets. In the cases of bad
names/realms there will not be an infinite loop, the function will
timeout on its own and return as an internal error.