Fix corruption resistance check on essences
Some essences have one and other have two spaces after the colon in "Unique:  Corruption..." (yay Blizz consistency) and because you already check for an exact string value why not directly search for its position instead of doing this risky split.
Also changed "totalStats[7] += 10;" to be ready for the case if the order of returned values in the API might change one day. The old version would overwrite the existing value instead of adding to it. Works currently because NECK is after/below the BACK item.